+++ src/main/java/com/devthink/devthink_server/application/ReviewService.java
@@ -2,17 +2,21 @@ package com.devthink.devthink_server.application;
 import com.devthink.devthink_server.domain.Book;
 import com.devthink.devthink_server.domain.Review;
 import com.devthink.devthink_server.domain.User;
+import com.devthink.devthink_server.dto.BookRequestData;
 import com.devthink.devthink_server.dto.ReviewDetailResponseData;
 import com.devthink.devthink_server.dto.ReviewRequestData;
 import com.devthink.devthink_server.dto.ReviewResponseData;
 import com.devthink.devthink_server.errors.ReviewNotFoundException;
+import com.devthink.devthink_server.errors.UserNotFoundException;
 import com.devthink.devthink_server.infra.BookRepository;
 import com.devthink.devthink_server.infra.ReviewRepository;
+import com.devthink.devthink_server.infra.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
+import java.util.Optional;
 
 @Service
 @Transactional(readOnly = true)
@@ -21,14 +25,18 @@ public class ReviewService {
 
     private final ReviewRepository reviewRepository;
     private final BookRepository bookRepository;
+    private final UserRepository userRepository;
 
     /**
      * 전달된 값으로 리뷰를 생성합니다.
-     * @param user, book, reviewRequestDto
+     * @param reviewRequestData
      * @return 생성 된 리뷰 id
      */
     @Transactional
-    public String createReview(User user, Book book, ReviewRequestData reviewRequestData){
+    public String createReview(ReviewRequestData reviewRequestData){
+        User user = userRepository.findByIdAndDeletedIsFalse(reviewRequestData.getUserId())
+                .orElseThrow(()-> new UserNotFoundException(reviewRequestData.getUserId()));
+        Book book = getBookByIsbn(reviewRequestData.getBook());
         Review review = reviewRepository.save(
             Review.builder()
                     .user(user)
@@ -40,6 +48,36 @@ public class ReviewService {
         review.getBook().addReview(review);
         review.getBook().setScoreAvg(bookRepository.calcScoreAvg(book.getId()));
         return review.getId().toString();
+    }
+
+    /**
+     * 입력된 isbn 정보로 Book을 조회하며, 해당 책이 없는 경우 새로 생성하는 함수를 호출합니다.
+     * @param bookRequestData (책에 대한 정보)
+     * @return 조회 혹은 생성된 Book 객체
+     */
+    public Book getBookByIsbn(BookRequestData bookRequestData) {
+        Optional<Book> book = bookRepository.getBookByIsbn(bookRequestData.getIsbn());
+        if (book.isEmpty()) {
+            return createBook(bookRequestData);
+        } else {
+            return book.get();
+        }
+    }
+
+    /**
+     * 입력된 책 정보의 새로운 Book 객체를 생성합니다.
+     * @param bookRequestData (책에 대한 정보)
+     * @return 생성된 Book 객체
+     */
+    @Transactional
+    public Book createBook(BookRequestData bookRequestData) {
+        Book book = Book.builder()
+                .isbn(bookRequestData.getIsbn())
+                .name(bookRequestData.getName())
+                .writer(bookRequestData.getWriter())
+                .imgUrl(bookRequestData.getImgUrl())
+                .build();
+        return bookRepository.save(book);
     }
 
     /**
